[rnd] Use the new links for mer-core under releases.nemomobile.org structure.
In the past the testing and devel always pointed to latest/next of mer core.
However while developing the devel is populated first with new release, and
the testing only when the devel is considered stable enough and all issues
have been fixed. Now we can control which version of mer-core is used
by ssu by users using rnd repositories.

Signed-off-by: Marko Saukko <marko.saukko@jollamobile.com>